fixed brandamage with the request_id, it used to sleep one second on every
Wed Sep 09 01:47:44 1998  George Lebl  <jirka@5z.com>

        * orbit-applet-lib.c: fixed brandamage with the request_id, it
          used to sleep one second on every call, and increase the timeouts
          to 30 secs

+++ gnome-panel/extern.c
@@ -332,7 +332,7 @@ reserve_applet_spot (Extern *ext, PanelWidget *panel, int pos,
 	if(!GTK_WIDGET_REALIZED(socket))
 		gtk_widget_realize(socket);
 
-	printf("socket wid: %lu\n", GDK_WINDOW_XWINDOW(socket->window));
+	printf("socket wid: %lu\n", (gulong)GDK_WINDOW_XWINDOW(socket->window));
 	return GDK_WINDOW_XWINDOW(socket->window);
 }
